Move RoleSelectionActivity to its own package
