+++ CloudyCar/app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
             android:name=".signup.SignUpActivity"
             android:label="@string/title_activity_main"
             android:theme="@style/AppTheme.NoActionBar"/>
-        <activity android:name=".RoleSelectionActivity"/>
+        <activity android:name=".roleselection.RoleSelectionActivity"/>
         <activity android:name=".createrequest.CreateRequestActivity"/>
         <activity
             android:name=".SummaryActivity"
