+++ CloudyCar/app/src/main/res/layout/activity_role_selection.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.cloudycrew.cloudycar.RoleSelectionActivity">
+    tools:context="com.cloudycrew.cloudycar.roleselection.RoleSelectionActivity">
 
     <android.support.v7.widget.CardView
         android:id="@+id/rider_card"
